requirements.txt
streamlit
pandas
numpy
google-generativeai
pdfplumber
yfinance
altair
openpyxl
cryptography
lxml